refactor: extract Plot3D into plot3d/ subpackage
Moves Plot3D and _triangulate_grid out of figure_plots.py into the new
anyplotlib/plot3d/ subpackage. A backward-compat shim in figure_plots.py
re-exports Plot3D so existing imports continue to work.
