+++ anyplotlib/plot3d/__init__.py
@@ -1,0 +1,3 @@
+from anyplotlib.plot3d._plot3d import Plot3D
+
+__all__ = ["Plot3D"]
